refactor: remove dump and refactor attention_backend
