+++ atom/plugin/rtpllm/attention_backend/__init__.py
@@ -1,4 +1,11 @@
-from .rtp_attention import RTPAttention
+from .attention_gdn import apply_attention_gdn_rtpllm_patch
+from .attention_switch import apply_attention_mha_rtpllm_patch
+from .rtp_full_attention import RTPAttention, RTPFullAttention
 
-__all__ = ["RTPAttention"]
+__all__ = [
+    "RTPAttention",
+    "RTPFullAttention",
+    "apply_attention_gdn_rtpllm_patch",
+    "apply_attention_mha_rtpllm_patch",
+]
