+++ atom/plugin/rtpllm/attention_backend/attention_switch.py
@@ -2,7 +2,7 @@ import logging
 
 from atom.plugin.prepare import is_rtpllm
 
-logger = logging.getLogger("atom.plugin.rtpllm.model_ops.attention_switch")
+logger = logging.getLogger("atom.plugin.rtpllm.attention_backend.attention_switch")
 
 _PATCHED = False
 _ORIGINAL_ATTENTION_CLS = None
@@ -16,17 +16,16 @@ def apply_attention_mha_rtpllm_patch() -> None:
         return
 
     import atom.model_ops as ops
-    from atom.plugin.rtpllm.attention_backend import RTPAttention
+    from .rtp_full_attention import RTPFullAttention
 
     if _ORIGINAL_ATTENTION_CLS is None:
         _ORIGINAL_ATTENTION_CLS = getattr(ops, "Attention", None)
 
     if is_rtpllm():
-        ops.RTPAttention = RTPAttention
-        ops.Attention = RTPAttention
+        ops.RTPFullAttention = RTPFullAttention
+        ops.Attention = RTPFullAttention
         logger.info(
-            "Applied RTP-LLM attention patch: atom.model_ops.Attention -> RTPAttention."
+            "Applied RTP-LLM attention patch: atom.model_ops.Attention -> RTPFullAttention."
         )
 
     _PATCHED = True
-
