--- atom/plugin/rtpllm/model_ops/__init__.py
@@ -1,4 +1,0 @@
-from .attention_gdn import apply_attention_gdn_rtpllm_patch
-from .attention_switch import apply_attention_mha_rtpllm_patch
-
-__all__ = ["apply_attention_gdn_rtpllm_patch", "apply_attention_mha_rtpllm_patch"]
